Re-organize (#9) and fix examples (#16)
